requirements.txt
# FastAPI y dependencias web
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0

# Base de datos y ORM  
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Pydantic para validación
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP cliente para CAPTCHA
aiohttp==3.9.1

# Templates y archivos estáticos
jinja2==3.1.2
python-multipart==0.0.6

# Utilidades
python-dotenv==1.0.0

# Producción
gunicorn==21.2.0